+++ mlxtend/frequent_patterns/fpmax.py
@@ -61,6 +61,11 @@ def fpmax(df, min_support=0.5, use_colnames=False, max_len=None, verbose=0):
     """
     fpc.valid_input_check(df)
 
+    if min_support <= 0.:
+        raise ValueError('`min_support` must be a positive '
+                         'number within the interval `(0, 1]`. '
+                         'Got %s.' % min_support)
+
     colname_map = None
     if use_colnames:
         colname_map = {idx: item for idx, item in enumerate(df.columns)}
@@ -78,6 +83,8 @@ def fpmax_step(tree, minsup, mfit, colnames, max_len, verbose):
     count = 0
     items = list(tree.nodes.keys())
     largest_set = sorted(tree.cond_items+items, key=mfit.rank.get)
+    if len(largest_set) == 0:
+        return
     if tree.is_path():
         if not mfit.contains(largest_set):
             count += 1
@@ -85,7 +92,7 @@ def fpmax_step(tree, minsup, mfit, colnames, max_len, verbose):
             mfit.cache = largest_set
             mfit.insert_itemset(largest_set)
             if max_len is None or len(largest_set) <= max_len:
-                support = min([tree.nodes[i][0].count for i in items])
+                support = tree.root.count
                 yield support, largest_set
 
     if verbose: