+++ mlxtend/frequent_patterns/tests/test_fpgrowth.py
@@ -1,15 +1,26 @@
 import unittest
 import numpy as np
-from mlxtend.frequent_patterns.tests.test_fpbase import FPTestAll
+from test_fpbase import FPTestEdgeCases, FPTestErrors, \
+                        FPTestEx1All, FPTestEx2All, FPTestEx3All
 from mlxtend.frequent_patterns import fpgrowth
 
 
-class TestFPGrowth(unittest.TestCase, FPTestAll):
+class TestEdgeCases(unittest.TestCase, FPTestEdgeCases):
     def setUp(self):
-        FPTestAll.setUp(self, fpgrowth)
+        FPTestEdgeCases.setUp(self, fpgrowth)
 
 
-class TestFPGrowth2(unittest.TestCase, FPTestAll):
+class TestErrors(unittest.TestCase, FPTestErrors):
+    def setUp(self):
+        FPTestErrors.setUp(self, fpgrowth)
+
+
+class TestEx1(unittest.TestCase, FPTestEx1All):
+    def setUp(self):
+        FPTestEx1All.setUp(self, fpgrowth)
+
+
+class TestEx1BoolInput(unittest.TestCase, FPTestEx1All):
     def setUp(self):
         one_ary = np.array(
             [[False, False, False, True, False, True, True, True, True,
@@ -22,4 +33,14 @@ class TestFPGrowth2(unittest.TestCase, FPTestAll):
               True, True],
              [False, True, False, True, True, True, False, False, True,
               False, False]])
-        FPTestAll.setUp(self, fpgrowth, one_ary=one_ary)
+        FPTestEx1All.setUp(self, fpgrowth, one_ary=one_ary)
+
+
+class TestEx2(unittest.TestCase, FPTestEx2All):
+    def setUp(self):
+        FPTestEx2All.setUp(self, fpgrowth)
+
+
+class TestEx3(unittest.TestCase, FPTestEx3All):
+    def setUp(self):
+        FPTestEx3All.setUp(self, fpgrowth)